+++ packages/vsce/src/commands/phaseInCommand.ts
@@ -77,7 +77,7 @@ export function getPhaseInCommand(tree: CICSTree, treeview: TreeView<any>) {
         if (!parentNode.getChildResource().resources.getFetchedAll()) {
           numToFetch -= 1;
         }
-        parentNode.getChildResource().resources.setNumberToFetch(numToFetch);
+        parentNode.setNumberToFetch(numToFetch);
 
         tree._onDidChangeTreeData.fire(parentNode);
       }